// frontend/src/model/movieResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * cinematch-api
 * OpenAPI spec version: 0.1.0
 */
import type { CastMemberResponse } from './castMemberResponse';

export interface MovieResponse {
  cast: CastMemberResponse[];
  /** @nullable */
  director?: string | null;
  genres: string[];
  /** @nullable */
  imdb_id?: string | null;
  /** @nullable */
  mediawiki_id?: string | null;
  movie_id: number;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  popularity?: number | null;
  /** @nullable */
  poster_url?: string | null;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  release_date?: string | null;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  tagline?: string | null;
  title: string;
  trailers: string[];
}